+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Release 1.5.0 (?)
 
 - Add clipboard copy funcionality for the full service instance id (#135)
+- Fix validation error for optional boolean inputs (#140)
 
 # Release 1.4.0 (2020-12-17)